# Conflicts:
#	holidaysomething/src/main/java/com/holidaysomething/holidaysomething/HolidaysomethingApplication.java
#	holidaysomething/src/main/java/com/holidaysomething/holidaysomething/repository/MemberRepository.java
#	holidaysomething/src/main/java/com/holidaysomething/holidaysomething/repository/custom/MemberRepositoryImpl.java
#	holidaysomething/src/main/java/com/holidaysomething/holidaysomething/security/SecurityConfig.java

+++ holidaysomething/src/main/java/com/holidaysomething/holidaysomething/domain/ProductImage.java
@@ -26,6 +26,7 @@ public class ProductImage {
 
   private Long category;
 
+// TODO : storedFileName에 경로도 같이 저장하면 불러오기가 편한데 사용한 곳이 있는 것 같다...
   private String path;
 
   @Column(nullable = false)
@@ -35,7 +36,7 @@ public class ProductImage {
   private String storedFileName;
 
   @Column(nullable = false)
-  private int size;
+  private Long size;
 
   @Column(nullable = false)
   private String fileType;
@@ -44,6 +45,7 @@ public class ProductImage {
   @CreationTimestamp
   private LocalDateTime regDate;
 
+//  TODO : 필요없어보이는데... 샘플 SQL을 고칠까??
   @UpdateTimestamp
   private LocalDateTime updateDate;
